Wire up the results to the users page
