+++ imports/ui/views/users/UsersList/UserListFilterBar.jsx
@@ -1,0 +1,9 @@
+import React, { PropTypes } from 'react'
+
+const UserListFilterBar = ({ }) => {
+  return (
+    <span>FILTERED!</span>
+  )
+}
+
+export default UserListFilterBar;
